# api/src/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PulseOps V1 API",
    "description": "Enterprise Modular Operations Platform — Backend API with hot-drop module architecture, JWT authentication, and role-based access control.",
    "version": "1.0.0",
    "contact": {
      "name": "PulseOps Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4001/api",
      "description": "Local Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /auth/login"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "admin@test.com" },
          "password": { "type": "string", "example": "Infosys@123" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "accessToken": { "type": "string" },
              "refreshToken": { "type": "string" },
              "expiresIn": { "type": "integer", "example": 86400 },
              "user": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "email": { "type": "string" },
                  "name": { "type": "string" },
                  "role": { "type": "string" }
                }
              }
            }
          }
        }
      },
      "ConfigEntry": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "value": { "type": "object" },
          "description": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and get JWT tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "description": "Invalid credentials" },
          "429": { "description": "Too many login attempts" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "New access token issued" },
          "401": { "description": "Invalid or expired refresh token" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout and invalidate session",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Logged out successfully" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user profile",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile returned" },
          "401": { "description": "Not authenticated" }
        }
      }
    },
    "/database/create-database": {
      "post": {
        "tags": ["Database"],
        "summary": "Create the application database if it does not exist",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Database created or already exists" },
          "500": { "description": "Creation failed" }
        }
      }
    },
    "/database/test-connection": {
      "get": {
        "tags": ["Database"],
        "summary": "Test database connection using server config",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Connection successful with latency" },
          "500": { "description": "Connection failed" }
        }
      }
    },
    "/database/schema-status": {
      "get": {
        "tags": ["Database"],
        "summary": "Check schema initialization status",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Schema status returned" }
        }
      }
    },
    "/database/create-schema": {
      "post": {
        "tags": ["Database"],
        "summary": "Create core schema and tables",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Schema created successfully" },
          "500": { "description": "Schema creation failed" }
        }
      }
    },
    "/database/load-default-data": {
      "post": {
        "tags": ["Database"],
        "summary": "Seed default admin user and core module",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Default data loaded" },
          "500": { "description": "Seeding failed" }
        }
      },
      "delete": {
        "tags": ["Database"],
        "summary": "Clean default seeded data",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Default data cleaned" },
          "500": { "description": "Clean failed" }
        }
      }
    },
    "/database/wipe": {
      "post": {
        "tags": ["Database"],
        "summary": "Drop entire schema (DESTRUCTIVE!)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Database wiped" },
          "500": { "description": "Wipe failed" }
        }
      }
    },
    "/database/stats": {
      "get": {
        "tags": ["Database"],
        "summary": "Get table sizes and counts",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Stats returned" }
        }
      }
    },
...
    "/config": {
      "get": {
        "tags": ["Config"],
        "summary": "List all configuration entries",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of config entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ConfigEntry" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Config"],
        "summary": "Create or update a configuration entry",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConfigEntry" }
            }
          }
        },
        "responses": {
          "200": { "description": "Config saved" },
          "400": { "description": "Key is required" }
        }
      }
    },
    "/config/{key}": {
      "get": {
        "tags": ["Config"],
        "summary": "Get a specific configuration entry",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Config entry" },
          "404": { "description": "Config not found" }
        }
      }
    }
  }
}